Abstract functions to pkg
