+++ core/manager/health_check.go
@@ -1,15 +1,11 @@
 package manager
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
-	"os/exec"
 	"path/filepath"
 	"strings"
 	"time"
 
-	"github.com/docker/docker/api/types"
 	"github.com/sdslabs/beastv4/core"
 	"github.com/sdslabs/beastv4/core/cache"
 	"github.com/sdslabs/beastv4/core/config"
@@ -236,238 +232,13 @@ func ProcessInstanceDeletionQueue() {
 func CleanupOrphanedInstanceContainers() {
 	log.Debug("Checking for orphaned instance containers")
 
-	cleanupOrphanedOnServer(core.LOCALHOST)
-	cleanupOrphanedComposeInstancesOnServer(core.LOCALHOST)
+	cr.CleanupOrphans()
+	cr.CleanupOrphanedComposeInstances()
 
 	for host, server := range config.Cfg.AvailableServers {
 		if server.Active && host != core.LOCALHOST {
-			cleanupOrphanedOnServer(host)
-			cleanupOrphanedComposeInstancesOnServer(host)
+			remoteManager.CleanupOrphanedOnServer(host)
+			remoteManager.CleanupOrphanedComposeInstancesOnServer(host)
 		}
 	}
-}
-
-func cleanupOrphanedOnServer(serverHost string) {
-	var containers []types.Container
-	var err error
-
-	if serverHost == core.LOCALHOST {
-		containers, err = cr.SearchContainerByFilter(map[string]string{
-			"label": "beast.instance=true",
-		})
-	} else {
-		server := config.Cfg.AvailableServers[serverHost]
-		containers, err = remoteManager.SearchContainerByFilterRemote(map[string]string{
-			"label": "beast.instance=true",
-		}, server)
-	}
-
-	if err != nil {
-		log.Warnf("Failed to search for instance containers on %s: %v", serverHost, err)
-		return
-	}
-
-	for _, container := range containers {
-		instanceID := container.Labels["beast.instance.id"]
-		if instanceID == "" {
-			for _, name := range container.Names {
-				name = strings.TrimPrefix(name, "/")
-				if strings.HasPrefix(name, "beast_instance_") {
-					parts := strings.Split(name, "_")
-					if len(parts) >= 4 {
-						instanceID = parts[len(parts)-1]
-						break
-					}
-				}
-			}
-		}
-
-		if instanceID == "" {
-			continue
-		}
-
-		_, err := cache.GetInstance(instanceID)
-		if err != nil {
-			containerName := ""
-			if len(container.Names) > 0 {
-				containerName = strings.TrimPrefix(container.Names[0], "/")
-			}
-			log.Infof("Removing orphaned instance container: %s (instance %s) on %s", containerName, instanceID, serverHost)
-
-			if serverHost == core.LOCALHOST {
-				if err := cr.StopAndRemoveContainer(container.ID); err != nil {
-					log.Warnf("Failed to remove orphaned container %s: %v", container.ID[:12], err)
-				}
-			} else {
-				server := config.Cfg.AvailableServers[serverHost]
-				if err := remoteManager.StopAndRemoveContainerRemote(container.ID, server); err != nil {
-					log.Warnf("Failed to remove orphaned container %s on %s: %v", container.ID[:12], serverHost, err)
-				}
-			}
-
-			cache.FreeContainerPortsOnHost(serverHost, container.ID)
-		}
-	}
-}
-
-// cleanupOrphanedComposeInstancesOnServer finds and removes orphaned docker compose instance projects.
-// Docker Compose containers don't have the beast.instance labels, but they have
-// com.docker.compose.project labels with project names starting with "beast-instance-".
-func cleanupOrphanedComposeInstancesOnServer(serverHost string) {
-	var projectNames []string
-	var err error
-
-	if serverHost == core.LOCALHOST {
-		projectNames, err = getOrphanedComposeInstanceProjects()
-	} else {
-		server := config.Cfg.AvailableServers[serverHost]
-		projectNames, err = getOrphanedComposeInstanceProjectsRemote(server)
-	}
-
-	if err != nil {
-		log.Warnf("Failed to get compose instance projects on %s: %v", serverHost, err)
-		return
-	}
-
-	for _, projectName := range projectNames {
-		// Extract instance ID from project name: beast-instance-{encoded_challenge}-{instanceID}
-		parts := strings.Split(projectName, "-")
-		if len(parts) < 4 {
-			continue
-		}
-		instanceID := parts[len(parts)-1]
-
-		// Check if instance still exists in cache
-		_, err := cache.GetInstance(instanceID)
-		if err != nil {
-			log.Infof("Removing orphaned compose instance project: %s (instance %s) on %s", projectName, instanceID, serverHost)
-
-			if serverHost == core.LOCALHOST {
-				if err := composeDownProject(projectName); err != nil {
-					log.Warnf("Failed to remove orphaned compose project %s: %v", projectName, err)
-				}
-			} else {
-				server := config.Cfg.AvailableServers[serverHost]
-				if err := composeDownProjectRemote(projectName, server); err != nil {
-					log.Warnf("Failed to remove orphaned compose project %s on %s: %v", projectName, serverHost, err)
-				}
-			}
-		}
-	}
-}
-
-// getOrphanedComposeInstanceProjects returns a list of docker compose project names
-// that match the instance naming pattern (beast-instance-*)
-func getOrphanedComposeInstanceProjects() ([]string, error) {
-	cmd := exec.Command("docker", "compose", "ls", "--format", "json")
-	var output bytes.Buffer
-	cmd.Stdout = &output
-	cmd.Stderr = &output
-
-	if err := cmd.Run(); err != nil {
-		return nil, fmt.Errorf("docker compose ls failed: %v, output: %s", err, output.String())
-	}
-
-	type ComposeProject struct {
-		Name   string `json:"Name"`
-		Status string `json:"Status"`
-	}
-
-	var projects []ComposeProject
-	outputStr := strings.TrimSpace(output.String())
-	if outputStr == "" {
-		return nil, nil
-	}
-
-	if err := json.Unmarshal([]byte(outputStr), &projects); err != nil {
-		// Try parsing line by line (older docker compose versions)
-		for _, line := range strings.Split(outputStr, "\n") {
-			if strings.TrimSpace(line) == "" {
-				continue
-			}
-			var project ComposeProject
-			if err := json.Unmarshal([]byte(line), &project); err != nil {
-				continue
-			}
-			projects = append(projects, project)
-		}
-	}
-
-	var instanceProjects []string
-	for _, project := range projects {
-		if strings.HasPrefix(project.Name, "beast-instance-") {
-			instanceProjects = append(instanceProjects, project.Name)
-		}
-	}
-
-	return instanceProjects, nil
-}
-
-// getOrphanedComposeInstanceProjectsRemote returns compose instance projects on a remote server
-func getOrphanedComposeInstanceProjectsRemote(server config.AvailableServer) ([]string, error) {
-	output, err := remoteManager.RunCommandOnServer(server, "docker compose ls --format json")
-	if err != nil {
-		return nil, fmt.Errorf("docker compose ls failed on remote: %v", err)
-	}
-
-	type ComposeProject struct {
-		Name   string `json:"Name"`
-		Status string `json:"Status"`
-	}
-
-	var projects []ComposeProject
-	outputStr := strings.TrimSpace(output)
-	if outputStr == "" {
-		return nil, nil
-	}
-
-	if err := json.Unmarshal([]byte(outputStr), &projects); err != nil {
-		// Try parsing line by line
-		for _, line := range strings.Split(outputStr, "\n") {
-			if strings.TrimSpace(line) == "" {
-				continue
-			}
-			var project ComposeProject
-			if err := json.Unmarshal([]byte(line), &project); err != nil {
-				continue
-			}
-			projects = append(projects, project)
-		}
-	}
-
-	var instanceProjects []string
-	for _, project := range projects {
-		if strings.HasPrefix(project.Name, "beast-instance-") {
-			instanceProjects = append(instanceProjects, project.Name)
-		}
-	}
-
-	return instanceProjects, nil
-}
-
-// composeDownProject removes a docker compose project by name
-func composeDownProject(projectName string) error {
-	cmd := exec.Command("docker", "compose", "-p", projectName, "down", "--remove-orphans", "-v")
-	var output bytes.Buffer
-	cmd.Stdout = &output
-	cmd.Stderr = &output
-
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("docker compose down failed: %v, output: %s", err, output.String())
-	}
-
-	log.Debugf("Successfully removed compose project %s", projectName)
-	return nil
-}
-
-// composeDownProjectRemote removes a docker compose project on a remote server
-func composeDownProjectRemote(projectName string, server config.AvailableServer) error {
-	cmd := fmt.Sprintf("docker compose -p %s down --remove-orphans -v", projectName)
-	output, err := remoteManager.RunCommandOnServer(server, cmd)
-	if err != nil {
-		return fmt.Errorf("docker compose down failed on remote: %v, output: %s", err, output)
-	}
-
-	log.Debugf("Successfully removed compose project %s on %s", projectName, server.Host)
-	return nil
 }
